+++ src/RavenNest.DataModels/NPCItemDrop.cs
@@ -7,6 +7,6 @@ namespace RavenNest.DataModels
         private Guid id; public Guid Id { get => id; set => Set(ref id, value); }
         private Guid npcId; public Guid NpcId { get => npcId; set => Set(ref npcId, value); }
         private Guid itemId; public Guid ItemId { get => itemId; set => Set(ref itemId, value); }
-        private decimal dropChance; public decimal DropChance { get => dropChance; set => Set(ref dropChance, value); }
+        private double dropChance; public double DropChance { get => dropChance; set => Set(ref dropChance, value); }
     }
 }
